# Day5/4_api-with-db-swagger/src/config/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Employee API",
        "description": "API documentation for employee management",
        "version": "1.0.0",
        "contact": {
            "email": "manish@ti.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local server"
        }
    ],
    "tags": [
        {
            "name": "Employees",
            "description": "Employee management"
        }
    ],
    "paths": {
        "/employees": {
            "get": {
                "summary": "List all employees",
                "tags": [
                    "Employees"
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of employees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Employee"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new employee",
                "tags": [
                    "Employees"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "eid": {
                                        "type": "integer"
                                    },
                                    "ename": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "eid",
                                    "ename"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Employee created successfully"
                    }
                }
            }
        },
        "/employees/{empid}": {
            "get": {
                "summary": "Get a single employee by ID",
                "tags": [
                    "Employees"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "empid",
                        "required": true,
                        "description": "ID of the employee",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the employee details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Employee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "put": {
                "summary": "Update an employee by ID",
                "tags": [
                    "Employees"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "empid",
                        "required": true,
                        "description": "ID of the employee",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "eid": {
                                        "type": "integer"
                                    },
                                    "ename": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Employee updated successfully"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an employee by ID",
                "tags": [
                    "Employees"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "empid",
                        "required": true,
                        "description": "ID of the employee to delete",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee deleted successfully"
                    },
                    "404": {
                        "description": "Employee not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Employee": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    }
                }
            }
        }
    }
}